requirements.txt
requests>=2.32.0
pandas>=2.2.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
beautifulsoup4>=4.12.3
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
